Create the LabelResolutionContext
Previously this context was merged inside LoweringContext. The downside is that label resolution was seen as a special pass, compared to the others. This commit extract label-resolution context out of lowering context, so that the lowering information is dropped at the end of the lowering step, instead of staying live until the end of the compilation.