Add `inline_*` macros and use it in parser tests
When writing parser test, it was super tedious to define the expected
AST. The resulting code was ugly and not obvious to the user. We deserve
better.

This commit introduces a whole bunch of macro whose goal is to parse
Rust tokens and produce the corresponding `dyl`-specific AST. It uses
only declarative macros and does most of the work at compile time. The
goal is to use it in order to be able to generate ASTs for the unit
tests all across the codebase.

The macro implementation is not trivial - almost unreadable for those
who don't know what to expect - but it was fun to write. It deserves a
blogpost for sure.